backend/pyproject.toml
[project]
name = "invest-report-backend"
version = "0.1.0"
description = "투자 레포팅 Agent 백엔드 API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "python-multipart>=0.0.6",
    "crewai>=0.28.0",
    "crewai-tools>=0.1.0",
    "finance-datareader>=0.9.50",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.3.23",
    "langchain-core>=0.3.83",
    "langgraph>=1.0.1",
]
